Move chunk manager to a seperate file
